requirements.txt
behave==1.2.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-openapi3==2.0.1
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.28.1
pydantic==1.9.2
psycopg2-binary==2.9.3
pytest==7.1.2
python-decouple==3.6
python-dotenv==0.20.0
requests==2.28.1
Werkzeug==2.2.2